Move use-action from utils to interactivity
