+++ packages/dara-core/js/shared/interactivity/index.tsx
@@ -5,6 +5,7 @@ export { useDataVariable } from './use-data-variable';
 export { combineFilters } from './filtering';
 export { default as useRefreshSelector } from './use-refresh-selector';
 export { default as useVariableState } from './use-variable-state';
+export { default as useAction, useActionIsLoading } from './use-action';
 export * from './store';
 export * from './use-server-component';
 export * from './nested';
