+++ packages/dara-core/js/shared/interactivity/persistence.tsx
@@ -8,7 +8,7 @@ import { validateResponse } from '@darajs/ui-utils';
 
 import { type BackendStorePatchMessage, type WebSocketClientInterface } from '@/api';
 import { RequestExtrasSerializable, request } from '@/api/http';
-import { handleAuthErrors } from '@/auth';
+import { handleAuthErrors } from '@/auth/auth';
 import { getSessionToken } from '@/auth/use-session-token';
 import { isEmbedded } from '@/shared/utils/embed';
 import { type GlobalTaskContext, type SingleVariable, isDerivedVariable } from '@/types';
